src/pyproject.toml
[tool.poetry]
name = "auth_example_with_fastapi"
version = "0.1.0"
description = ""
authors = ["fealone <fealone@lonesec.com>"]

[tool.poetry.dependencies]
python = "^3.7"
bcrypt = "3.2.0"
cffi = "1.14.4"
click = "7.0"
dnspython = "2.0.0"
email-validator = "1.1.2"
fastapi = "0.45.0"
h11 = "0.9.0"
httptools = "0.0.13"
idna = "2.10"
passlib = "1.7.4"
pycparser = "2.20"
pydantic = "1.3"
pymysql = "0.10.1"
six = "1.15.0"
sqlalchemy = "1.3.20"
starlette = "0.12.9"
typing-extensions = "3.7.4.3"
uvicorn = "0.13.1"
uvloop = "0.14.0"
websockets = "8.1"
PyJWT = "^1.7.1"
cryptography = "^3.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
